1. Single non-rigid deformation to keep things simple and prevent a lot of images from folding 2. Change padding_mode in the torch.nn.functional.grid_sample() to 'border' to avoid artifacts at image boundaries after resampling

+++ src/util_scripts/deformations.py
@@ -52,19 +52,20 @@ def create_bspline_transform(coarse=True,
     x, y, z = shape
 
     if coarse is True:
-        random_grid = np.random.rand(3, 4, 4, 4).astype(np.float32) # Make a random 3D 3 x 3 x 3 grid
+        random_grid = np.random.rand(3, 3, 3, 3).astype(np.float32) # Make a random 3D 3 x 3 x 3 grid
         random_grid = random_grid*2 - 1 # Shift range to [-1, 1]
         random_grid[0, ...] = random_grid[0, ...]*(2/x)*displacements[0]
         random_grid[1, ...] = random_grid[1, ...]*(2/y)*displacements[1]
         random_grid[2, ...] = random_grid[2, ...]*(2/z)*displacements[2]
     else: # Fine with a larger control grid
-        random_grid = np.random.rand(3, 8, 8, 8).astype(np.float32) # Make a random 3D 3 x 3 x 3 grid
+        random_grid = np.random.rand(3, 8, 8, 8).astype(np.float32) # Make a random 3D 4 x 4 x 4 grid
         random_grid = random_grid*2 - 1 # Shift range to [-1, 1]
         random_grid[0, ...] = random_grid[0, ...]*(2/x)*displacements[0]
         random_grid[1, ...] = random_grid[1, ...]*(2/y)*displacements[1]
         random_grid[2, ...] = random_grid[2, ...]*(2/z)*displacements[2]
 
-    bspline_transform = gryds.BSplineTransformation(random_grid)
+    bspline_transform = gryds.BSplineTransformation(grid=random_grid,
+                                                    order=3)
 
     return bspline_transform
 
@@ -113,7 +114,9 @@ def create_batch_deformation_grid(shape,
                                   device='cpu',
                                   dummy=False,
                                   non_rigid=True,
-                                  coarse_displacements=(1.5, 1.5, 1.5),
+                                  coarse=True,
+                                  fine=False,
+                                  coarse_displacements=(4, 4, 3),
                                   fine_displacements=(0.75, 0.75, 0.75)):
 
     b, c, i, j, k = shape
@@ -127,24 +130,23 @@ def create_batch_deformation_grid(shape,
 
     # Loop over batch and generated a unique deformation grid for each image in the batch
     for batch_idx in range(b):
-
+        transforms = []
         if dummy is False:
             if non_rigid is True:
-                elastic_transform_coarse = create_bspline_transform(coarse=True,
-                                                                    shape=[k, j, i],
-                                                                    displacements=coarse_displacements)
+                if coarse is True:
+                    elastic_transform_coarse = create_bspline_transform(coarse=True,
+                                                                        shape=[k, j, i],
+                                                                        displacements=coarse_displacements)
+                    transforms.append(elastic_transform_coarse)
+                if fine is True:
+                    elastic_transform_fine = create_bspline_transform(coarse=False,
+                                                                      shape=[k, j, i],
+                                                                      displacements=fine_displacements)
+                    transforms.append(elastic_transform_fine)
 
-                elastic_transform_fine = create_bspline_transform(coarse=False,
-                                                                  shape=[k, j, i],
-                                                                  displacements=fine_displacements)
-
-                transforms = [elastic_transform_coarse, elastic_transform_fine]
             else: # Translation only
                 translation_transform = create_affine_transform(translation=[0, 0, 0.1])
-                transforms = [translation_transform]
-        else: # No transforms
-            transforms = []
-
+                transforms.append(translation_transform)
 
         # Create deformation grid by composing transforms
         deformed_grid = create_deformation_grid(shape=[k, j, i],
@@ -163,9 +165,9 @@ def create_batch_deformation_grid(shape,
 
         batch_deformation_grid[batch_idx, ...] = deformed_grid
 
-    #Deform the whole batch by stacking all deformation grids along the batch axis (dim=0)
-    batch_deformation_grid = torch.Tensor(batch_deformation_grid).to(device)
-    return batch_deformation_grid
+        #Deform the whole batch by stacking all deformation grids along the batch axis (dim=0)
+        batch_deformation_grid = torch.Tensor(batch_deformation_grid).to(device)
+        return batch_deformation_grid
 
 
 
@@ -177,9 +179,10 @@ if __name__ == '__main__':
     train_dict = create_data_dicts_lesion_matching([train_patients[0]])
 
     data_loader, transforms = create_dataloader_lesion_matching(data_dicts=train_dict,
-                                                                train=False,
+                                                                train=True,
                                                                 batch_size=1,
-                                                                data_aug=False)
+                                                                data_aug=True,
+                                                                patch_size=(96, 96, 48))
 
     post_transforms = Compose([EnsureTyped(keys=['d_image']),
                                Invertd(keys=['d_image',
@@ -199,38 +202,42 @@ if __name__ == '__main__':
         images = batch_data['image']
         batch_deformation_grid = create_batch_deformation_grid(shape=images.shape,
                                                                dummy=False,
-                                                               non_rigid=True)
-        deformed_images = F.grid_sample(input=images,
-                                        grid=batch_deformation_grid,
-                                        align_corners=True,
-                                        mode="bilinear")
+                                                               non_rigid=True,
+                                                               coarse=True,
+                                                               fine=False)
+        if batch_deformation_grid is not None:
+            deformed_images = F.grid_sample(input=images,
+                                            grid=batch_deformation_grid,
+                                            align_corners=True,
+                                            mode="bilinear",
+                                            padding_mode="border")
 
-        sub_image = images - deformed_images
+            sub_image = images - deformed_images
 
-        save_dir = 'images_translation_b_{}'.format(b_id)
+            save_dir = 'images_translation_b_{}'.format(b_id)
 
-        if os.path.exists(save_dir) is True:
-            shutil.rmtree(save_dir)
+            if os.path.exists(save_dir) is True:
+                shutil.rmtree(save_dir)
 
-        os.makedirs(save_dir)
+            os.makedirs(save_dir)
 
-        for batch_idx in range(images.shape[0]):
-            image = images[batch_idx, ...]
-            dimage = deformed_images[batch_idx, ...]
-            image_array = torch.squeeze(image, dim=0).numpy() # Get rid of channel axis
-            dimage_array = torch.squeeze(dimage, dim=0).numpy()
-            image_fname = os.path.join(save_dir, 'image.nii')
-            dimage_fname = os.path.join(save_dir, 'deformed_image.nii')
-            image_nib = create_nibabel_image(image_array=image_array,
-                                             metadata_dict=batch_data['image_meta_dict'],
-                                             affine=batch_data['image_meta_dict']['affine'][batch_idx])
+            for batch_idx in range(images.shape[0]):
+                image = images[batch_idx, ...]
+                dimage = deformed_images[batch_idx, ...]
+                image_array = torch.squeeze(image, dim=0).numpy() # Get rid of channel axis
+                dimage_array = torch.squeeze(dimage, dim=0).numpy()
+                image_fname = os.path.join(save_dir, 'image.nii')
+                dimage_fname = os.path.join(save_dir, 'deformed_image.nii')
+                image_nib = create_nibabel_image(image_array=image_array,
+                                                 metadata_dict=batch_data['image_meta_dict'],
+                                                 affine=batch_data['image_meta_dict']['affine'][batch_idx])
 
-            dimage_nib = create_nibabel_image(image_array=dimage_array,
-                                              metadata_dict=batch_data['image_meta_dict'],
-                                              affine=batch_data['image_meta_dict']['affine'][batch_idx])
+                dimage_nib = create_nibabel_image(image_array=dimage_array,
+                                                  metadata_dict=batch_data['image_meta_dict'],
+                                                  affine=batch_data['image_meta_dict']['affine'][batch_idx])
 
-            save_nib_image(image_nib,
-                           image_fname)
+                save_nib_image(image_nib,
+                               image_fname)
 
-            save_nib_image(dimage_nib,
-                           dimage_fname)
+                save_nib_image(dimage_nib,
+                               dimage_fname)
